+++ spec/support/engine_context.rb
@@ -58,7 +58,7 @@ RSpec.shared_context "with mounted llm cost tracker engine" do
     create_lct_tables!
     LlmCostTracker::Call.reset_column_information
     LlmCostTracker::ServiceCharge.reset_column_information
-    LlmCostTracker::PeriodTotal.reset_column_information
+    LlmCostTracker::CallRollup.reset_column_information
     LlmCostTracker::Ingestion::InboxEntry.reset_column_information
     LlmCostTracker::Ingestion::Lease.reset_column_information
   end